+++ src/model/mod.rs
@@ -3,5 +3,6 @@ pub mod comment;
 pub mod listing;
 pub mod post;
 pub mod preview;
+pub mod responses;
 pub mod sort_time;
 pub mod token;