+++ src/model/responses/comment_response.rs
@@ -1,0 +1,12 @@
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct CommentResponse {
+    pub json: Option<ErrorResponse>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ErrorResponse {
+    pub errors: Vec<Vec<String>>,
+}