+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add examples [ `best_posts.rs` , `top_posts.rs` ]
 - Add Generics to Listings. Can use as `Listing<Post>` or `Listing<Comment>`
+- Add comment endpoint for threads
 
 ## 0.1.1 (2020/15/07)